requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic==2.8.2
openai>=1.40.0
requests>=2.32.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
